progress on the dataloader of fastformer
