+++ src/ebrec/utils/utils_behaviors.py
@@ -93,7 +93,7 @@ def create_binary_labels_column(
     _COLUMNS = df.columns
     GROUPBY_ID = generate_unique_name(_COLUMNS, "_groupby_id")
 
-    df = df.with_row_count(GROUPBY_ID)
+    df = df.with_row_index(GROUPBY_ID)
 
     if shuffle:
         df = shuffle_list_column(df, column=inview_col, seed=seed)
@@ -114,13 +114,12 @@ def create_binary_labels_column(
 
 
 def create_user_id_mapping(
-    df: pl.DataFrame,
-    user_col: str = DEFAULT_USER_COL,
+    df: pl.DataFrame, user_col: str = DEFAULT_USER_COL, value_str: str = "id"
 ):
     return create_lookup_dict(
-        df.select(pl.col(user_col).unique()).with_row_count("id"),
+        df.select(pl.col(user_col).unique()).with_row_index(value_str),
         key=user_col,
-        value="id",
+        value=value_str,
     )
