--- test/__init__.py
@@ -1,6 +1,0 @@
-from pytest.test_dataloader import (
-    test_FastformerDataloader,
-    test_LSTURDataLoader,
-    test_NAMLDataLoader,
-    test_NRMSDataLoader,
-)
