+++ test/dataloader/test_dataloader_newsrec.py
@@ -24,6 +24,9 @@ from ebrec.utils.constants import (
     DEFAULT_USER_COL,
 )
 
+from ebrec.models.fastformer.dataloader import FastformerDataset
+from torch.utils.data import DataLoader
+
 N_SAMPLES = "n"
 TOKEN_COL = "idx"
 BATCH_SIZE = 100
@@ -95,7 +98,7 @@ def test_NRMSDataLoader():
     test_dataloader = NRMSDataLoader(
         behaviors=df_behaviors,
         article_dict=article_mapping,
-        history_column=HISTORY_COLUMN,
+        history_column=DEFAULT_HISTORY_ARTICLE_ID_COL,
         unknown_representation="zeros",
         eval_mode=True,
         batch_size=BATCH_SIZE,
@@ -190,3 +193,47 @@ def test_NAMLDataLoader():
     assert isinstance(
         batch[1].ravel()[0], np.integer
     ), "Expected output to be integer; this is label"
+
+
+@time_it(True)
+def test_FastformerDataloader():
+
+    train_dataloader = DataLoader(
+        FastformerDataset(
+            behaviors=df_behaviors_train,
+            history_column=DEFAULT_HISTORY_ARTICLE_ID_COL,
+            article_dict=article_mapping,
+            batch_size=BATCH_SIZE,
+            shuffle=True,
+        )
+    )
+
+    batch = train_dataloader.__iter__().__next__()
+
+    assert train_dataloader.__len__() == int(np.ceil(df_behaviors_train.shape[0] / 100))
+    assert len(batch) == 2, "There should be two outputs: (inputs, labels)"
+    assert (
+        len(batch[0]) == 2
+    ), "Fastformer has two outputs (history_input, candidate_input)"
+
+    for type_in_batch in batch[0]:
+        assert (
+            type_in_batch.dtype == torch.int
+        ), "Expected output to be integer; used for lookup value"
+
+    assert batch[1].dtype == torch.float, "Expected output to be integer; this is label"
+
+    test_dataloader = DataLoader(
+        FastformerDataset(
+            behaviors=df_behaviors,
+            history_column=DEFAULT_HISTORY_ARTICLE_ID_COL,
+            article_dict=article_mapping,
+            batch_size=BATCH_SIZE,
+            shuffle=False,
+        )
+    )
+
+    batch = test_dataloader.__iter__().__next__()
+    assert len(batch[1].squeeze(0)) == sum(
+        label_lengths[:BATCH_SIZE]
+    ), "Should have unfolded all the test samples"
